change the color scheme for coef/spec plots
red / grey / blue for neg / insig / pos

I did this because i noticed in some of the specification plots where all effects are sig and in the same direction, it's hard to quickly tell what that direction is because you have to look for a small gray line or the axis for that.

Also synced the paper and SI color schemes in this regard.